+++ app/javascript/packs/application.js
@@ -8,6 +8,7 @@ require("@rails/ujs").start()
 require("@rails/activestorage").start()
 require("channels")
 require("../map")
+require("../get_my_place")
 
 
 // Uncomment to copy all static images under ../images to the output folder and reference
